requirements.txt
fastapi==0.110.1
pydantic-core==2.16.3
uvicorn==0.29.0
pytest==8.1.1
httpx==0.27.0
pydantic_settings==2.2.1
sqlalchemy==2.0.29
asyncpg==0.29.0
psycopg2==2.9.9
aioredis==2.0.1
alembic==1.13.1